refactor: give the STK push API a name that isn't about one caller
The public STK push endpoints were named for Klik POS and lived in
api/sales_invoice.py, so the path read
api.sales_invoice.initiate_klik_pos_stk_push. Neither half was true: any
external system can use them, and they are not about sales invoices.

Move them to api/stk_push.py as create_stk_push_request,
get_stk_push_status and link_stk_push_to_document. The link endpoint now
takes reference_name/reference_doctype rather than invoice_name/
invoice_doctype, since the document need not be an invoice.

Breaking: callers must move to the new paths. Klik POS is updated in its
own repo and the two must deploy together.
